Moved classes into subnamespaces
