+++ src/Generators/ManifestGenerator.php
@@ -9,7 +9,10 @@
  * file that was distributed with this source code.
  */
 
-namespace Fusonic\WebApp;
+namespace Fusonic\WebApp\Generators;
+
+use Fusonic\WebApp\AppConfiguration;
+use Fusonic\WebApp\Objects\Image;
 
 /**
  * Generates an app manifest according to the W3C specification "Manifest for a web application"
